complete Wufoo ajax request
We're making a cross-origin POST. Wufoo will accept our request, but the
browser doesn't believe the request is successful. The downside here is
that we can't confirm the submission was actually successful.

Two types of errors are likely:
- We've hit our rate limit
- A valid email wasn't provided

We can account for the second item using HTML5 input types or other UX
validations. We'd just have to count requests for the first item.