scale height of info box on mobile
Info boxes default to a minimum height of 240px. This allows for
consistent readability if the page is full-width and the text doesn't
aggressively wrap.

When necessary, allow height to scale to fully display all of the
written content. This could potentially cause the two boxes to scale to
different heights. However, on smaller screens, the boxes will already
move to stack on top of each other, so this difference won't be
noticeable.