redirect /docs to /docs/basic-setup
For the cleanest user experience, we'll make /basic-setup it's own page,
and directly link all top-level references to Documentation to that
page.

An alternative method is to make the content of /index.html match our
basic-setup page. Eventually, /index.html could become a landing page
that links to multiple locations.

With the chosen approach, we create a common-sense page hierarchy,
support direct links to /docs by the user, and support web crawling that
Googlebot and others will do.