adjust canonical url to www. subdomain
Using GitHub Pages as an apex domain was causing issues with the domain
redirection. Specifically, sometimes GitHub's servers would return a 302
response (back to itself) instead of returning the website. This is
believed to be part of the Denial of Service protection.

This redirection prevents some services from easily visiting the
website. Specifically: Facebook's open graph scraper. This impacts the
ability to share links from this domain.